Added option to skip existing wordlists and --force to force overwrite.

+++ Content Management System (CMS)/croogo/gen_croogo_wordlists.py
@@ -15,7 +15,9 @@ from bs4 import BeautifulSoup
 
 def parseArgs():
     parser = argparse.ArgumentParser(description="Description message")
-    parser.add_argument("-v", "--verbose", default=None, action="store_true", help='arg1 help message')
+    parser.add_argument("-v", "--verbose", default=None, action="store_true", help='Verbose mode (default: False)')
+    parser.add_argument("-f", "--force", default=None, action="store_true", help='Force updating existing wordlists. (default: False)')
+    parser.add_argument("-n", "--no-commit", default=False, action="store_true", help='Disable automatic commit (default: False)')
     return parser.parse_args()
 
 
@@ -44,43 +46,56 @@ if __name__ == '__main__':
         if "sig" not in a['href']:
             version = a['href'].split('/v')[1].rstrip('.zip')
             croogo_versions[version] = "https://downloads.croogo.org/" + a['href']
+    print('[>] Loaded %d croogo versions.' % len(croogo_versions.keys()))
 
     for version in sorted(croogo_versions.keys()):
-        print('   [>] Extracting wordlist of croogo version %s ...' % version)
+        str_version = version
 
-        dl_url = croogo_versions[version]
-        if not os.path.exists('./versions/%s/' % version):
-            os.makedirs('./versions/%s/' % version, exist_ok=True)
+        generate = False
+        if not os.path.exists('./versions/%s/' % str_version):
+            os.makedirs('./versions/%s/' % str_version, exist_ok=True)
+            generate = True
+        elif options.force:
+            generate = True
+        elif options.verbose:
+            print('[>] Ignoring croogo version %s (local wordlists exists)' % str_version)
 
-        if options.verbose:
-            print("      [>] Create dir ...")
-        os.system('rm -rf /tmp/paths_croogo_extract/; mkdir -p /tmp/paths_croogo_extract/')
-        if options.verbose:
-            print("      [>] Getting file ...")
-            print('wget -q --show-progress "%s" -O /tmp/paths_croogo_extract/croogo.zip' % dl_url)
-            os.system('wget -q --show-progress "%s" -O /tmp/paths_croogo_extract/croogo.zip' % dl_url)
-        else:
-            os.popen('wget -q "%s" -O /tmp/paths_croogo_extract/croogo.zip' % dl_url).read()
-        if options.verbose:
-            print("      [>] Unzipping archive ...")
-        os.system('cd /tmp/paths_croogo_extract/; unzip croogo.zip 1>/dev/null; rm croogo.zip')
+        if generate:
+            print('[>] Extracting wordlists for croogo version %s' % str_version)
 
-        if options.verbose:
-            print("      [>] Getting wordlist ...")
-        if version in ["1.1"]:
-            save_wordlist(os.popen('cd /tmp/paths_croogo_extract/; find .').read(), version, filename="croogo.txt")
-            save_wordlist(os.popen('cd /tmp/paths_croogo_extract/; find . -type f').read(), version, filename="croogo_files.txt")
-            save_wordlist(os.popen('cd /tmp/paths_croogo_extract/; find . -type d').read(), version, filename="croogo_dirs.txt")
-        else:
-            save_wordlist(os.popen('cd /tmp/paths_croogo_extract/croogo*/; find .').read(), version, filename="croogo.txt")
-            save_wordlist(os.popen('cd /tmp/paths_croogo_extract/croogo*/; find . -type f').read(), version, filename="croogo_files.txt")
-            save_wordlist(os.popen('cd /tmp/paths_croogo_extract/croogo*/; find . -type d').read(), version, filename="croogo_dirs.txt")
+            dl_url = croogo_versions[version]
+            if not os.path.exists('./versions/%s/' % version):
+                os.makedirs('./versions/%s/' % version, exist_ok=True)
 
-        if options.verbose:
-            print("      [>] Committing results ...")
-            os.system('git add ./versions/%s/; git commit -m "Added wordlists for croogo version %s";' % (version, version))
-        else:
-            os.popen('git add ./versions/%s/; git commit -m "Added wordlists for croogo version %s";' % (version, version)).read()
+            if options.verbose:
+                print("      [>] Create dir ...")
+            os.system('rm -rf /tmp/paths_croogo_extract/; mkdir -p /tmp/paths_croogo_extract/')
+            if options.verbose:
+                print("      [>] Getting file ...")
+                print('wget -q --show-progress "%s" -O /tmp/paths_croogo_extract/croogo.zip' % dl_url)
+                os.system('wget -q --show-progress "%s" -O /tmp/paths_croogo_extract/croogo.zip' % dl_url)
+            else:
+                os.popen('wget -q "%s" -O /tmp/paths_croogo_extract/croogo.zip' % dl_url).read()
+            if options.verbose:
+                print("      [>] Unzipping archive ...")
+            os.system('cd /tmp/paths_croogo_extract/; unzip croogo.zip 1>/dev/null; rm croogo.zip')
+
+            if options.verbose:
+                print("      [>] Getting wordlist ...")
+            if version in ["1.1"]:
+                save_wordlist(os.popen('cd /tmp/paths_croogo_extract/; find .').read(), version, filename="croogo.txt")
+                save_wordlist(os.popen('cd /tmp/paths_croogo_extract/; find . -type f').read(), version, filename="croogo_files.txt")
+                save_wordlist(os.popen('cd /tmp/paths_croogo_extract/; find . -type d').read(), version, filename="croogo_dirs.txt")
+            else:
+                save_wordlist(os.popen('cd /tmp/paths_croogo_extract/croogo*/; find .').read(), version, filename="croogo.txt")
+                save_wordlist(os.popen('cd /tmp/paths_croogo_extract/croogo*/; find . -type f').read(), version, filename="croogo_files.txt")
+                save_wordlist(os.popen('cd /tmp/paths_croogo_extract/croogo*/; find . -type d').read(), version, filename="croogo_dirs.txt")
+
+            if options.verbose:
+                print("      [>] Committing results ...")
+                os.system('git add ./versions/%s/; git commit -m "Added wordlists for croogo version %s";' % (version, version))
+            else:
+                os.popen('git add ./versions/%s/; git commit -m "Added wordlists for croogo version %s";' % (version, version)).read()
 
     if options.verbose:
         print("      [>] Creating common wordlists ...")
