+++ Content Management System (CMS)/lavalite/gen_lavalite_wordlists.py
@@ -15,8 +15,11 @@ from bs4 import BeautifulSoup
 
 def parseArgs():
     parser = argparse.ArgumentParser(description="Description message")
-    parser.add_argument("-v", "--verbose", default=None, action="store_true", help='arg1 help message')
+    parser.add_argument("-v", "--verbose", default=None, action="store_true", help='Verbose mode (default: False)')
+    parser.add_argument("-f", "--force", default=None, action="store_true", help='Force updating existing wordlists. (default: False)')
+    parser.add_argument("-n", "--no-commit", default=False, action="store_true", help='Disable automatic commit (default: False)')
     return parser.parse_args()
+
 
 def get_releases_from_github(username, repo, per_page=100):
     # https://docs.github.com/en/rest/reference/repos#releases
@@ -62,41 +65,50 @@ if __name__ == '__main__':
     versions = get_releases_from_github("lavalite", "cms")
 
     for version in versions.keys():
-        print('[>] Extracting wordlist for lavalite version %s' % version)
+        str_version = version
 
-        if not os.path.exists('./versions/%s/' % version):
-            os.makedirs('./versions/%s/' % version, exist_ok=True)
+        generate = False
+        if not os.path.exists('./versions/%s/' % (str_version)):
+            os.makedirs('./versions/%s/' % (str_version), exist_ok=True)
+            generate = True
+        elif options.force:
+            generate = True
+        elif options.verbose:
+            print('[>] Ignoring lavalite version %s (local wordlists exists)' % str_version)
 
-        dl_url = versions[version]
+        if generate:
+            print('[>] Extracting wordlists for lavalite version %s' % str_version)
 
-        if options.verbose:
-            print("      [>] Create dir ...")
-            os.system('rm -rf /tmp/paths_lavalite_extract/; mkdir -p /tmp/paths_lavalite_extract/')
-        else:
-            os.popen('rm -rf /tmp/paths_lavalite_extract/; mkdir -p /tmp/paths_lavalite_extract/').read()
-        if options.verbose:
-            print("      [>] Getting file ...")
-            print('wget -q --show-progress "%s" -O /tmp/paths_lavalite_extract/lavalite.zip' % dl_url)
-            os.system('wget -q --show-progress "%s" -O /tmp/paths_lavalite_extract/lavalite.zip' % dl_url)
-        else:
-            os.popen('wget -q "%s" -O /tmp/paths_lavalite_extract/lavalite.zip' % dl_url).read()
-        if options.verbose:
-            print("      [>] Unzipping archive ...")
-            os.system('cd /tmp/paths_lavalite_extract/; unzip lavalite.zip 1>/dev/null')
-        else:
-            os.popen('cd /tmp/paths_lavalite_extract/; unzip lavalite.zip 1>/dev/null').read()
+            dl_url = versions[version]
 
-        if options.verbose:
-            print("      [>] Getting wordlist ...")
-        save_wordlist(os.popen('cd /tmp/paths_lavalite_extract/*/; find .').read(), version, filename="lavalite.txt")
-        save_wordlist(os.popen('cd /tmp/paths_lavalite_extract/*/; find . -type f').read(), version, filename="lavalite_files.txt")
-        save_wordlist(os.popen('cd /tmp/paths_lavalite_extract/*/; find . -type d').read(), version, filename="lavalite_dirs.txt")
+            if options.verbose:
+                print("      [>] Create dir ...")
+                os.system('rm -rf /tmp/paths_lavalite_extract/; mkdir -p /tmp/paths_lavalite_extract/')
+            else:
+                os.popen('rm -rf /tmp/paths_lavalite_extract/; mkdir -p /tmp/paths_lavalite_extract/').read()
+            if options.verbose:
+                print("      [>] Getting file ...")
+                print('wget -q --show-progress "%s" -O /tmp/paths_lavalite_extract/lavalite.zip' % dl_url)
+                os.system('wget -q --show-progress "%s" -O /tmp/paths_lavalite_extract/lavalite.zip' % dl_url)
+            else:
+                os.popen('wget -q "%s" -O /tmp/paths_lavalite_extract/lavalite.zip' % dl_url).read()
+            if options.verbose:
+                print("      [>] Unzipping archive ...")
+                os.system('cd /tmp/paths_lavalite_extract/; unzip lavalite.zip 1>/dev/null')
+            else:
+                os.popen('cd /tmp/paths_lavalite_extract/; unzip lavalite.zip 1>/dev/null').read()
 
-        if options.verbose:
-            print("      [>] Committing results ...")
-            os.system('git add ./versions/%s/*; git commit -m "Added wordlists for lavalite version %s";' % (version, version))
-        else:
-            os.popen('git add ./versions/%s/*; git commit -m "Added wordlists for lavalite version %s";' % (version, version)).read()
+            if options.verbose:
+                print("      [>] Getting wordlist ...")
+            save_wordlist(os.popen('cd /tmp/paths_lavalite_extract/*/; find .').read(), version, filename="lavalite.txt")
+            save_wordlist(os.popen('cd /tmp/paths_lavalite_extract/*/; find . -type f').read(), version, filename="lavalite_files.txt")
+            save_wordlist(os.popen('cd /tmp/paths_lavalite_extract/*/; find . -type d').read(), version, filename="lavalite_dirs.txt")
+
+            if options.verbose:
+                print("      [>] Committing results ...")
+                os.system('git add ./versions/%s/*; git commit -m "Added wordlists for lavalite version %s";' % (version, version))
+            else:
+                os.popen('git add ./versions/%s/*; git commit -m "Added wordlists for lavalite version %s";' % (version, version)).read()
 
     if options.verbose:
         print("      [>] Creating common wordlists ...")
