+++ Content Management System (CMS)/wordpress/gen_wordpress_wordlists.py
@@ -15,7 +15,9 @@ from bs4 import BeautifulSoup
 
 def parseArgs():
     parser = argparse.ArgumentParser(description="Description message")
-    parser.add_argument("-v", "--verbose", default=None, action="store_true", help='arg1 help message')
+    parser.add_argument("-v", "--verbose", default=None, action="store_true", help='Verbose mode (default: False)')
+    parser.add_argument("-f", "--force", default=None, action="store_true", help='Force updating existing wordlists. (default: False)')
+    parser.add_argument("-n", "--no-commit", default=False, action="store_true", help='Disable automatic commit (default: False)')
     return parser.parse_args()
 
 
@@ -77,43 +79,51 @@ if __name__ == '__main__':
                     versions[version][lang] = dllink
 
     for version in versions.keys():
-        print('[>] Using wordpress version %s' % version)
 
         for lang in versions[version].keys():
             print('   [>] Extracting wordlist of version %s lang %s ...' % (version, lang))
 
+            generate = False
             if not os.path.exists('./versions/%s/%s/' % (version, lang)):
                 os.makedirs('./versions/%s/%s/' % (version, lang), exist_ok=True)
+                generate = True
+            elif options.force:
+                generate = True
+            elif options.verbose:
+                print('[>] Ignoring wordpress version %s-%s (local wordlists exists)' % (version, lang))
 
-            dl_url = versions[version][lang]
+            if generate:
+                print('[>] Extracting wordlists for wordpress version %s-%s' % (version, lang))
 
-            if options.verbose:
-                print("      [>] Create dir ...")
-                os.system('rm -rf /tmp/paths_wordpress_extract/; mkdir -p /tmp/paths_wordpress_extract/')
-            else:
-                os.popen('rm -rf /tmp/paths_wordpress_extract/; mkdir -p /tmp/paths_wordpress_extract/').read()
-            if options.verbose:
-                print("      [>] Getting file ...")
-                os.system('wget -q --show-progress "%s" -O /tmp/paths_wordpress_extract/wordpress.zip' % dl_url)
-            else:
-                os.popen('wget -q "%s" -O /tmp/paths_wordpress_extract/wordpress.zip' % dl_url).read()
-            if options.verbose:
-                print("      [>] Unzipping archive ...")
-                os.system('cd /tmp/paths_wordpress_extract/; unzip wordpress.zip 1>/dev/null')
-            else:
-                os.popen('cd /tmp/paths_wordpress_extract/; unzip wordpress.zip 1>/dev/null').read()
+                dl_url = versions[version][lang]
 
-            if options.verbose:
-                print("      [>] Getting wordlist ...")
-            save_wordlist(os.popen('cd /tmp/paths_wordpress_extract/wordpress/; find .').read(), version, lang, filename="wordpress.txt")
-            save_wordlist(os.popen('cd /tmp/paths_wordpress_extract/wordpress/; find . -type f').read(), version, lang, filename="wordpress_files.txt")
-            save_wordlist(os.popen('cd /tmp/paths_wordpress_extract/wordpress/; find . -type d').read(), version, lang, filename="wordpress_dirs.txt")
+                if options.verbose:
+                    print("      [>] Create dir ...")
+                    os.system('rm -rf /tmp/paths_wordpress_extract/; mkdir -p /tmp/paths_wordpress_extract/')
+                else:
+                    os.popen('rm -rf /tmp/paths_wordpress_extract/; mkdir -p /tmp/paths_wordpress_extract/').read()
+                if options.verbose:
+                    print("      [>] Getting file ...")
+                    os.system('wget -q --show-progress "%s" -O /tmp/paths_wordpress_extract/wordpress.zip' % dl_url)
+                else:
+                    os.popen('wget -q "%s" -O /tmp/paths_wordpress_extract/wordpress.zip' % dl_url).read()
+                if options.verbose:
+                    print("      [>] Unzipping archive ...")
+                    os.system('cd /tmp/paths_wordpress_extract/; unzip wordpress.zip 1>/dev/null')
+                else:
+                    os.popen('cd /tmp/paths_wordpress_extract/; unzip wordpress.zip 1>/dev/null').read()
 
-            if options.verbose:
-                print("      [>] Committing results ...")
-                os.system('git add ./versions/%s/%s/*; git commit -m "Added wordlists for wordpress version %s lang %s";' % (version, lang, version, lang))
-            else:
-                os.popen('git add ./versions/%s/%s/*; git commit -m "Added wordlists for wordpress version %s lang %s";' % (version, lang, version, lang)).read()
+                if options.verbose:
+                    print("      [>] Getting wordlist ...")
+                save_wordlist(os.popen('cd /tmp/paths_wordpress_extract/wordpress/; find .').read(), version, lang, filename="wordpress.txt")
+                save_wordlist(os.popen('cd /tmp/paths_wordpress_extract/wordpress/; find . -type f').read(), version, lang, filename="wordpress_files.txt")
+                save_wordlist(os.popen('cd /tmp/paths_wordpress_extract/wordpress/; find . -type d').read(), version, lang, filename="wordpress_dirs.txt")
+
+                if options.verbose:
+                    print("      [>] Committing results ...")
+                    os.system('git add ./versions/%s/%s/*; git commit -m "Added wordlists for wordpress version %s lang %s";' % (version, lang, version, lang))
+                else:
+                    os.popen('git add ./versions/%s/%s/*; git commit -m "Added wordlists for wordpress version %s lang %s";' % (version, lang, version, lang)).read()
 
     if options.verbose:
         print("      [>] Creating common wordlists ...")
